Moved examples into runnable form. Nearly completed simulation, inference, and strategy code.
